+++ gwcelery/conf/__init__.py
@@ -143,6 +143,8 @@ check_vector_prepost = {'gstlal': [2, 2],
                         'HardwareInjection': [2, 2],
                         'Swift': [2, 2],
                         'Fermi': [2, 2],
+                        'INTEGRAL': [2, 2],
+                        'AGILE': [2, 2],
                         'SNEWS': [10, 10]}
 """Seconds before and after the superevent start and end times which the DQ
 vector check will include in its check. Pipeline dependent."""
@@ -157,6 +159,8 @@ uses_gatedhoft = {'gstlal': True,
                   'HardwareInjection': False,
                   'Swift': False,
                   'Fermi': False,
+                  'INTEGRAL': False,
+                  'AGILE': False,
                   'SNEWS': False}
 """Whether or not a pipeline uses gated h(t). Determines whether or not
 the DMT-DQ_VECTOR will be analyzed for data quality."""
@@ -194,6 +198,8 @@ idq_veto = {'gstlal': False,
             'HardwareInjection': False,
             'Swift': False,
             'Fermi': False,
+            'INTEGRAL': False,
+            'AGILE': False,
             'SNEWS': False}
 """If true for a pipeline, iDQ values above the threshold defined in
 :obj:`~gwcelery.conf.idq_pglitch.thres` will cause DQV to be labeled.